# [ADD] product: Added index on field `product_id` of `product.price.history`